requirements.txt
fastapi
uvicorn
kafka-python
pyspark
apache-airflow
google-cloud-bigquery
google-cloud-storage
psycopg2-binary
pandas
faker
python-dotenv
sqlalchemy
pydantic